refactor: 패키지 구조를 레이어 아키텍쳐로 변경
