+++ src/main/java/roomescape/RoomEscapeWebApplication.java
@@ -1,4 +1,4 @@
-package web;
+package roomescape;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
