+++ src/main/java/roomescape/dto/response/ErrorResponseDTO.java
@@ -1,4 +1,4 @@
-package web.dto.response;
+package roomescape.dto.response;
 
 import lombok.AccessLevel;
 import lombok.Getter;
