+++ src/main/java/roomescape/repository/ReservationJdbcTemplateRepository.java
@@ -1,0 +1,5 @@
+package roomescape.repository;
+
+public class ReservationJdbcTemplateRepository {
+
+}
