+++ src/main/java/roomescape/repository/ReservationRepository.java
@@ -1,0 +1,5 @@
+package roomescape.repository;
+
+public interface ReservationRepository {
+
+}
