+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,0 +1,5 @@
+package roomescape.service;
+
+public class ReservationService {
+
+}
