+++ src/test/java/roomescape/controller/ReservationControllerTest.java
@@ -1,4 +1,4 @@
-package web.presentation;
+package roomescape.controller;
 
 import static org.hamcrest.core.Is.is;
 
@@ -14,9 +14,9 @@ import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.ActiveProfiles;
-import web.dto.request.ReservationRequestDTO;
-import web.exception.DuplicatedReservationException;
-import web.exception.NoSuchReservationException;
+import roomescape.dto.request.ReservationRequestDTO;
+import roomescape.exception.DuplicatedReservationException;
+import roomescape.exception.NoSuchReservationException;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
